cl: blockCtx typeParamLookup
